server/requirements.txt
openai==1.45.0
chromadb==0.5.5
PyPDF2==3.0.1
sentence-transformers==3.0.1
python-dotenv==1.0.1
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
typing-extensions>=4.0.0
google-generativeai==0.8.2

# LangChain 의존성 추가
langchain==0.1.0
langchain-core==0.1.13
langchain-openai==0.0.5
langchain-google-genai==0.0.7
langchain-community==0.0.13

# LangGraph 추가
langgraph==0.0.26
